bug fixes for ubuntu compiler

+++ cgg02_transform_viewing/Scene2D.cpp
@@ -62,7 +62,7 @@ Object2D* Scene2D::getObject(int index)
 }
 
 
-size_t Scene2D::getObjectSize()
+unsigned int Scene2D::getObjectSize()
 {
 	return objects_.size();
 }
